+++ tecnicaturagit3Py/Archivos/Leccion01/NumerosIgualesExcepciones.py
@@ -1,0 +1,3 @@
+class NumerosIgualesExcepciones (Exception): # EXTIENDE DE LA CLASE
+    def __init__(self, mensaje):
+        self.message = mensaje
