+++ tecnicaturagit3Py/Archivos/Leccion01/__pycache__/NumerosIgualesExcepciones.cpython-39.pyc
