--- tecnicaturagit3Py/excepciones/Leccion01/NumerosIgualesExcepciones.py
@@ -1,3 +1,0 @@
-class NumerosIgualesExcepciones (Exception): # EXTIENDE DE LA CLASE
-    def __init__(self, mensaje):
-        self.message = mensaje
